requirements.txt
python-telegram-bot==20.7
openai==1.12.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth==2.29.0
python-dotenv==1.0.0


